requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Cors==4.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytest==7.4.0
flask-swagger-ui==4.11.1
flask-wtf==1.2.1
